Refactor: 통합 DTO로 패키지경로 변경
