+++ src/main/java/sheetplus/checkings/business/auth/service/AuthService.java
@@ -6,7 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sheetplus.checkings.business.auth.dto.LoginDto;
-import sheetplus.checkings.domain.member.dto.response.MemberInfoResponseDto;
+import sheetplus.checkings.domain.member.dto.MemberDto.MemberInfoResponseDto;
 import sheetplus.checkings.domain.token.dto.TokenDto;
 import sheetplus.checkings.domain.member.entity.Member;
 import sheetplus.checkings.domain.enums.MemberType;
