+++ src/main/java/sheetplus/checkings/business/page/admin/service/AdminPageService.java
@@ -5,13 +5,13 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import sheetplus.checkings.business.page.admin.dto.AdminHomeResponseDto;
+import sheetplus.checkings.business.page.admin.dto.AdminPageDto.AdminHomeResponseDto;
 import sheetplus.checkings.domain.contest.entity.Contest;
 import sheetplus.checkings.domain.event.entity.Event;
-import sheetplus.checkings.domain.entry.dto.response.EntryInfoResponseDto;
-import sheetplus.checkings.domain.entry.dto.response.EntryResponseDto;
-import sheetplus.checkings.domain.event.dto.response.EventResponseDto;
-import sheetplus.checkings.domain.member.dto.response.MemberInfoResponseDto;
+import sheetplus.checkings.domain.entry.dto.EntryDto.EntryInfoResponseDto;
+import sheetplus.checkings.domain.entry.dto.EntryDto.EntryResponseDto;
+import sheetplus.checkings.domain.event.dto.EventDto.EventResponseDto;
+import sheetplus.checkings.domain.member.dto.MemberDto.MemberInfoResponseDto;
 import sheetplus.checkings.domain.contest.repository.ContestRepository;
 import sheetplus.checkings.domain.entry.repository.EntryRepository;
 import sheetplus.checkings.domain.member.repository.MemberRepository;
