+++ src/main/java/sheetplus/checkings/business/page/student/service/StudentPageService.java
@@ -5,9 +5,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import sheetplus.checkings.business.page.student.dto.ActivitiesResponseDto;
-import sheetplus.checkings.domain.event.dto.response.EventResponseDto;
-import sheetplus.checkings.business.page.student.dto.StudentHomePageResponseDto;
+import sheetplus.checkings.business.page.student.dto.StudentPageDto.ActivitiesResponseDto;
+import sheetplus.checkings.business.page.student.dto.StudentPageDto.StudentHomePageResponseDto;
+import sheetplus.checkings.domain.event.dto.EventDto.EventResponseDto;
 import sheetplus.checkings.domain.member.entity.Member;
 import sheetplus.checkings.domain.participatecontest.entity.ParticipateContest;
 import sheetplus.checkings.domain.enums.EventCategory;
