+++ src/main/java/sheetplus/checkings/business/page/superadmin/service/SuperAdminService.java
@@ -4,8 +4,8 @@ package sheetplus.checkings.business.page.superadmin.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import sheetplus.checkings.domain.member.dto.response.MemberInfoResponseDto;
-import sheetplus.checkings.domain.member.dto.request.MemberRequestDto;
+import sheetplus.checkings.domain.member.dto.MemberDto.MemberInfoResponseDto;
+import sheetplus.checkings.domain.member.dto.MemberDto.MemberRequestDto;
 import sheetplus.checkings.domain.token.dto.TokenDto;
 import sheetplus.checkings.domain.adminacceptcons.entity.AdminAcceptCons;
 import sheetplus.checkings.domain.member.entity.Member;
