+++ src/main/java/sheetplus/checkings/domain/contest/service/ContestCRUDService.java
@@ -5,8 +5,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import sheetplus.checkings.domain.contest.dto.request.ContestRequestDto;
-import sheetplus.checkings.domain.contest.dto.response.ContestResponseDto;
+import sheetplus.checkings.domain.contest.dto.ContestDto.ContestRequestDto;
+import sheetplus.checkings.domain.contest.dto.ContestDto.ContestResponseDto;
 import sheetplus.checkings.domain.contest.entity.Contest;
 import sheetplus.checkings.domain.contest.repository.ContestRepository;
 import sheetplus.checkings.exception.exceptionMethod.ApiException;
