+++ src/main/java/sheetplus/checkings/domain/draw/service/DrawEventService.java
@@ -4,10 +4,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import sheetplus.checkings.domain.draw.dto.request.DrawEventRequestDto;
-import sheetplus.checkings.domain.draw.dto.request.DrawUpdateRequestDto;
-import sheetplus.checkings.domain.draw.dto.response.DrawEventResponseDto;
-import sheetplus.checkings.domain.draw.dto.response.DrawUpdateResponseDto;
+import sheetplus.checkings.domain.draw.dto.DrawDto.DrawEventRequestDto;
+import sheetplus.checkings.domain.draw.dto.DrawDto.DrawUpdateRequestDto;
+import sheetplus.checkings.domain.draw.dto.DrawDto.DrawEventResponseDto;
+import sheetplus.checkings.domain.draw.dto.DrawDto.DrawUpdateResponseDto;
 import sheetplus.checkings.domain.contest.entity.Contest;
 import sheetplus.checkings.domain.draw.entity.Draw;
 import sheetplus.checkings.domain.member.entity.Member;
