+++ src/main/java/sheetplus/checkings/domain/entry/service/EntryCRUDService.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sheetplus.checkings.business.page.admin.controller.AdminPageController;
 import sheetplus.checkings.domain.entry.controller.EntryController;
-import sheetplus.checkings.domain.entry.dto.request.EntryRequestDto;
-import sheetplus.checkings.domain.entry.dto.response.EntryResponseDto;
+import sheetplus.checkings.domain.entry.dto.EntryDto.EntryRequestDto;
+import sheetplus.checkings.domain.entry.dto.EntryDto.EntryResponseDto;
 import sheetplus.checkings.domain.contest.entity.Contest;
 import sheetplus.checkings.domain.entry.entity.Entry;
 import sheetplus.checkings.domain.contest.repository.ContestRepository;
