+++ src/main/java/sheetplus/checkings/domain/event/service/EventCRUDService.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sheetplus.checkings.business.page.admin.controller.AdminPageController;
 import sheetplus.checkings.domain.event.controller.EventController;
-import sheetplus.checkings.domain.event.dto.request.EventRequestDto;
-import sheetplus.checkings.domain.event.dto.response.EventResponseDto;
+import sheetplus.checkings.domain.event.dto.EventDto.EventRequestDto;
+import sheetplus.checkings.domain.event.dto.EventDto.EventResponseDto;
 import sheetplus.checkings.domain.contest.entity.Contest;
 import sheetplus.checkings.domain.event.entity.Event;
 import sheetplus.checkings.domain.contest.repository.ContestRepository;
