+++ src/main/java/sheetplus/checkings/domain/favorite/service/FavoriteCRUDService.java
@@ -18,9 +18,9 @@ import sheetplus.checkings.domain.contest.repository.ContestRepository;
 import sheetplus.checkings.domain.event.entity.Event;
 import sheetplus.checkings.domain.event.repository.EventRepository;
 import sheetplus.checkings.domain.favorite.controller.FavoriteController;
-import sheetplus.checkings.domain.favorite.dto.request.FavoriteRequestDto;
-import sheetplus.checkings.domain.favorite.dto.response.FavoriteCreateResponseDto;
-import sheetplus.checkings.domain.favorite.dto.response.FavoriteResponseDto;
+import sheetplus.checkings.domain.favorite.dto.FavoriteDto.FavoriteRequestDto;
+import sheetplus.checkings.domain.favorite.dto.FavoriteDto.FavoriteCreateResponseDto;
+import sheetplus.checkings.domain.favorite.dto.FavoriteDto.FavoriteResponseDto;
 import sheetplus.checkings.domain.favorite.entity.Favorite;
 import sheetplus.checkings.domain.favorite.repository.FavoriteRepository;
 import sheetplus.checkings.domain.member.entity.Member;
