+++ src/main/java/sheetplus/checkings/domain/member/controller/MemberController.java
@@ -5,8 +5,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import sheetplus.checkings.business.auth.dto.LoginDto;
-import sheetplus.checkings.domain.member.dto.request.MemberRequestDto;
-import sheetplus.checkings.domain.member.dto.request.MemberUpdateRequestDto;
+import sheetplus.checkings.domain.member.dto.MemberDto.MemberRequestDto;
+import sheetplus.checkings.domain.member.dto.MemberDto.MemberUpdateRequestDto;
 import sheetplus.checkings.domain.token.dto.TokenDto;
 import sheetplus.checkings.domain.member.entity.Member;
 import sheetplus.checkings.business.auth.service.AuthService;
