+++ src/main/java/sheetplus/checkings/domain/member/service/MemberCRUDService.java
@@ -3,8 +3,8 @@ package sheetplus.checkings.domain.member.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import sheetplus.checkings.domain.member.dto.request.MemberRequestDto;
-import sheetplus.checkings.domain.member.dto.request.MemberUpdateRequestDto;
+import sheetplus.checkings.domain.member.dto.MemberDto.MemberRequestDto;
+import sheetplus.checkings.domain.member.dto.MemberDto.MemberUpdateRequestDto;
 import sheetplus.checkings.domain.member.entity.Member;
 import sheetplus.checkings.domain.adminacceptcons.repository.AdminAcceptConsRepository;
 import sheetplus.checkings.domain.member.repository.MemberRepository;
